At Ben's recommendation, move new ArgSymbol/FnSymbol methods to codegen/symbol
